refactor: 重複した内容を共通化
